feat(#302): Refactor term handling by creating separate classes for 'ordering' terms
