+++ .gitignore
@@ -9,3 +9,4 @@ node_modules/
 rdoc/
 vendor/
 .claude/
+.aidy/
