+++ test/factbase/terms/test_prev.rb
@@ -1,0 +1,22 @@
+# frozen_string_literal: true
+
+# SPDX-FileCopyrightText: Copyright (c) 2024-2025 Yegor Bugayenko
+# SPDX-License-Identifier: MIT
+
+# Test for unique term.
+# Author:: Volodya Lombrozo (volodya.lombrozo@gmail.com)
+# Copyright:: Copyright (c) 2024-2025 Yegor Bugayenko
+# License:: MIT
+
+require_relative '../../test__helper'
+require_relative '../../../lib/factbase/term'
+require_relative '../../../lib/factbase/terms/prev'
+
+class TestPrev < Factbase::Test
+  def test_prev
+    t = Factbase::Prev.new([:foo])
+    assert_nil(t.evaluate(fact('foo' => 41), [], Factbase.new))
+    assert_equal([41], t.evaluate(fact('foo' => 5), [], Factbase.new))
+    assert_equal([5], t.evaluate(fact('foo' => 6), [], Factbase.new))
+  end
+end
